dev/scripts/Python/requirements.txt
requests~=2.28.2
faker~=18.4.0
psycopg2-binary~=2.9.6
